edge_app/requirements_edge.txt
ultralytics>=8.1.0
opencv-python>=4.8.0
pyyaml>=6.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
